requirements.txt
pyairtable
google-generativeai
requests
python-dotenv